plus-btn-popup open and close behaviour and select-divider-list now done via state
