+++ .idea/inspectionProfiles/Project_Default.xml
@@ -4,8 +4,9 @@
     <inspection_tool class="HtmlUnknownAttribute" enabled="true" level="WARNING" enabled_by_default="true">
       <option name="myValues">
         <value>
-          <list size="1">
+          <list size="2">
             <item index="0" class="java.lang.String" itemvalue="movingup" />
+            <item index="1" class="java.lang.String" itemvalue="pop-up-state" />
           </list>
         </value>
       </option>
