+++ client/flash/src/components/css/HomePage.css
@@ -2,7 +2,7 @@
     width: 100%;
     grid-column: 2/ span 1;
     height: 100vh;
-    overflow-y: scroll;
+    overflow-y: auto;
 
 }
